Split Class_Schedule file into main.py and Schedule
