+++ Class_Schedule.py
@@ -1,57 +1,16 @@
 from __future__ import print_function
 from typing import List, Tuple, Dict, Any, Union
 from ortools.sat.python import cp_model
-import gspread
-from oauth2client.service_account import ServiceAccountCredentials
-from pathlib import Path
 from datetime import datetime, timedelta, date
 import itertools
-
 from ortools.sat.python.cp_model import IntVar, CpModel
-
 import pandas as pd
-import time
-import pickle as pickle
-import sys
 import re
-
-from Class_League import League, Club, Dates, Date, Team, CourtSlot, write_gsheet_output_data, get_gsheet_data
+from Class_League import League, write_gsheet_output_data, get_gsheet_data
 
 
 def main():
-    league_management_url = "https://docs.google.com/spreadsheets/d/1Mi-fWF63mw8Sdcb_lzHHTTvPqCp0VcJyZaqD5cm6D8U"
-    predefined_fixtures_url = "https://docs.google.com/spreadsheets/d/1kcnj4X01u5wpCghAi1yDCEBR54bnGv69bx-hVXVljIk"
-
-    league_2021 = reload_league_data_from_gsheet(_load_from_gsheets=False,
-                                                 _league_management_url=league_management_url)
-    # Print League data stats
-    league_2021.check_league_data()
-
-    # league_2021.write_teams_entered()
-
-    # league_2021.write_output()
-
-    schedule_2021 = Schedule(league_2021, predefined_fixtures_url)
-
-
-def reload_league_data_from_gsheet(_load_from_gsheets: bool, _league_management_url):
-    if _load_from_gsheets:
-        sys.setrecursionlimit(100000)
-        _league_2021 = League(_league_management_url)
-        with open('league2021.pkl', 'ab'):
-            pass
-        with open('league2021.pkl', 'wb') as f:
-            pickle.dump(_league_2021, f)
-        print("Session Saved")
-    else:
-        with open('league2021.pkl', 'rb') as f:
-            _league_2021 = pickle.load(f)
-        print("Session loaded")
-
-    if _league_2021:
-        return _league_2021
-    else:
-        print("Load Error")
+    pass
 
 
 class Schedule:
